requirements.txt
sentry-sdk>=1.28.1
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
python-dateutil>=2.8.2